pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "fx-ultimat-bot"
version = "1.0.0"
description = "Adaptive Multi-Dimensional AI Trading Bot for XAUUSDm (HFT/Scalping)"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "stable-baselines3>=2.2.0",
    "gymnasium>=0.29.0",
    "pandas>=2.1.0",
    "numpy>=1.25.0",
    "MetaTrader5>=5.0.45",
    "PyYAML>=6.0",
    "scikit-learn>=1.3.0",
    "ta>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "backtest*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
